requirements.txt
html2text==2020.1.16
bs4==0.0.1
lxml==4.6.2
pandas==1.2.4
beautifulsoup4==4.11.1
fpdf2==2.5.4
